Move generic diff functions to utils
